Make QueryParameterNameComparer internal.
