+++ src/Components/Components/src/Routing/QueryParameterValueSupplier.cs
@@ -187,19 +187,5 @@ namespace Microsoft.AspNetCore.Components.Routing
                 IsArray = isArray;
             }
         }
-
-        private class QueryParameterNameComparer : IComparer<ReadOnlyMemory<char>>, IEqualityComparer<ReadOnlyMemory<char>>
-        {
-            public static readonly QueryParameterNameComparer Instance = new();
-
-            public int Compare(ReadOnlyMemory<char> x, ReadOnlyMemory<char> y)
-                => x.Span.CompareTo(y.Span, StringComparison.OrdinalIgnoreCase);
-
-            public bool Equals(ReadOnlyMemory<char> x, ReadOnlyMemory<char> y)
-                => x.Span.Equals(y.Span, StringComparison.OrdinalIgnoreCase);
-
-            public int GetHashCode([DisallowNull] ReadOnlyMemory<char> obj)
-                => string.GetHashCode(obj.Span, StringComparison.OrdinalIgnoreCase);
-        }
     }
 }
